Housekeeping: Split AuthContextProvider into context, provider and hook.
