+++ frontend/src/Components/Users/Auth/Logout.tsx
@@ -1,7 +1,7 @@
 import { notification, Row, Spin } from 'antd';
 import React, { ReactElement, useEffect } from 'react';
 import { useNavigate } from 'react-router';
-import { useAuth } from '../../../Hooks/AuthContextProvider';
+import { useAuth } from '../../../Hooks/useAuth';
 
 export const Logout = (): ReactElement => {
   const auth = useAuth();
@@ -9,7 +9,7 @@ export const Logout = (): ReactElement => {
 
   useEffect(() => {
     async function ensureLogoutAndForwardToLogin () {
-      if (auth.user) {
+      if (auth && auth.user) {
         if (await auth.logout()) {
           notification.success({
             message: 'Successfully logged out.'
@@ -27,7 +27,7 @@ export const Logout = (): ReactElement => {
 
   return (
     <Row justify='center' align='middle' style={{ minHeight: '100vh', backgroundColor: '#001529' }}>
-      {auth.user && <Spin tip={`Logging out "${auth.user.name}"...`} />}
+      {auth && auth.user && <Spin tip={`Logging out "${auth.user.name}"...`} />}
     </Row>
   );
 };
