+++ frontend/src/Hooks/useAuth.ts
@@ -1,0 +1,6 @@
+import React from 'react';
+import { AuthContext, AuthContextType } from '../Contexts/AuthContext/AuthContext';
+
+export function useAuth (): AuthContextType | null {
+  return React.useContext(AuthContext);
+}
